refactor: extract lib.types package to decouple backend from frontend libs
Create a new zero-dependency lib.types package containing all shared
permission types, constants, and default configurations. This breaks the
problematic dependency chain where service.backend imported from
lib.permissions, which depends on lib.api (a frontend HTTP client).

Architecture:
- lib.types: pure types + data, zero deps, safe for both frontend + backend
- lib.permissions: re-exports lib.types + adds frontend service classes
- service.backend: imports from lib.types only (no frontend deps)

Changes:
- Create lib.types with RBAC types, field permission types, rescue
  permission constants, and default field permission configurations
- Update lib.permissions to re-export from lib.types (backwards compatible)
- Update service.backend to depend on lib.types instead of lib.permissions
- Update Dockerfile to build lib.types instead of lib.permissions + lib.api
- Update docker-compose.yml volume mounts
- Update CI path filters to trigger backend builds on lib.types changes
- Add lib.types to workspace configuration
