+++ lib.permissions/jest.config.cjs
@@ -35,6 +35,7 @@ module.exports = {
   "testTimeout": 10000,
   "moduleNameMapper": {
     "^@adopt-dont-shop/lib-api$": "<rootDir>/../lib.api/src/index.ts",
-    "^@adopt-dont-shop/lib\\.api$": "<rootDir>/../lib.api/src/index.ts"
+    "^@adopt-dont-shop/lib\\.api$": "<rootDir>/../lib.api/src/index.ts",
+    "^@adopt-dont-shop/lib\\.types$": "<rootDir>/../lib.types/src/index.ts"
   }
 };
