+++ lib.types/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@adopt-dont-shop/lib.permissions",
+  "name": "@adopt-dont-shop/lib.types",
   "version": "1.0.0",
-  "description": "Role-based access control and permissions management",
+  "description": "Shared TypeScript types, permission constants, and default configurations. Zero dependencies - safe for both frontend and backend.",
   "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -33,20 +33,18 @@
     "pet-adoption",
     "library",
     "typescript",
-    "react"
+    "types"
   ],
   "author": "Adopt Don't Shop Team",
   "license": "MIT",
   "dependencies": {
-    "@types/node": "^20.0.0",
-    "@adopt-dont-shop/lib.api": "file:../lib.api",
-    "@adopt-dont-shop/lib.types": "file:../lib.types"
+    "@types/node": "^20.0.0"
   },
   "devDependencies": {
     "@adopt-dont-shop/eslint-config-base": "*",
+    "@types/jest": "^29.4.0",
     "@typescript-eslint/eslint-plugin": "^7.0.0",
     "@typescript-eslint/parser": "^7.0.0",
-    "@types/jest": "^29.4.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
@@ -62,10 +60,10 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ParagonJenko/pet-adoption-react.git",
-    "directory": "lib.permissions"
+    "directory": "lib.types"
   },
   "bugs": {
     "url": "https://github.com/ParagonJenko/pet-adoption-react/issues"
   },
-  "homepage": "https://github.com/ParagonJenko/pet-adoption-react/tree/main/lib.permissions#readme"
+  "homepage": "https://github.com/ParagonJenko/pet-adoption-react/tree/main/lib.types#readme"
 }
