+++ lib.types/src/index.ts
@@ -1,0 +1,15 @@
+// Core RBAC types
+export * from './types';
+
+// Rescue-specific permission constants
+export * from './types/rescue-permissions';
+
+// Field-level permission types
+export * from './types/field-permissions';
+
+// Field permission default configurations
+export {
+  defaultFieldPermissions,
+  getDefaultFieldAccess,
+  getFieldAccessMap,
+} from './config/field-permission-defaults';
