+++ service.backend/package.json
@@ -30,7 +30,7 @@
     "predeploy": "npm run lint && npm run test && npm run security:validate && npm run build"
   },
   "dependencies": {
-    "@adopt-dont-shop/lib.permissions": "*",
+    "@adopt-dont-shop/lib.types": "file:../lib.types",
     "@sentry/node": "^10.23.0",
     "@sentry/profiling-node": "^10.23.0",
     "@types/bcrypt": "^5.0.0",
